+++ NEWS.md
@@ -11,7 +11,7 @@
 
 - Recurse into return types, so that for example `Result<bool>` can generate
   `Ok(true)` and `Ok(false)`, and `Some<T>` generates `None` and every generated
-  value of `T`. Similarly for `Box<T>`, and `Vec<T>`.
+  value of `T`. Similarly for `Box<T>`, `Vec<T>`, `Rc<T>`, `Arc<T>`.
 
 - Generate specific values for integers: `[0, 1]` for unsigned integers,
   `[0, 1, -1]` for signed integers; `[1]` for NonZero unsigned integers and