+++ NEWS.md
@@ -1,5 +1,10 @@
 # cargo-mutants changelog
 
+## Unreleased
+
+- Fixed a bug causing an assertion failure when cargo-mutants was run from a
+  subdirectory of a workspace.
+
 ## 23.6.0
 
 - Generate `Box::leak(Box::new(...))` as a mutation of functions returning